Convierte IndicePanel en enum class

+++ src/vista/vista_basics.cpp
@@ -2,5 +2,5 @@
 
 int obtener_posicion_x_panel(IndicePanel indice_panel) {
     return medidas::MARGEN_IZQ_PANELES +
-           (medidas::DESPLAZAMIENTO_LATERAL * indice_panel);
+           (medidas::DESPLAZAMIENTO_LATERAL * static_cast<int>(indice_panel));
 }
